requirements.txt
# Core Framework
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.5.0
pydantic-settings>=2.1.0

# JWT Authentication
PyJWT>=2.8.0
python-jose[cryptography]>=3.3.0
httpx>=0.25.0

# LangGraph and LangChain
langgraph>=0.0.40
langchain>=0.1.0
langchain-core>=0.1.0
langchain-community>=0.0.10

# MCP Support
mcp>=0.9.0

# Data Storage
pyyaml>=6.0.1
python-dotenv>=1.0.0

# Async Support
aiofiles>=23.2.1
aiohttp>=3.9.0

# Logging and Monitoring
structlog>=23.2.0

# OpenTelemetry
opentelemetry-api>=1.21.0
opentelemetry-sdk>=1.21.0
opentelemetry-exporter-otlp>=1.21.0
opentelemetry-instrumentation-fastapi>=0.42b0

# Development
pytest>=7.4.0
pytest-asyncio>=0.21.0
